add prompt hacking protection
