+++ lib/prompts/quality-check.ts
@@ -1,7 +1,7 @@
 import { LetterRequest } from '@/types/letter';
+import { QUALITY_CHECK_CRITERIA } from '../constants/ai';
 import { getPlatformPolicy, getRelevantPolicies } from '../platform-policies';
 import { platforms } from '../platforms';
-import { QUALITY_CHECK_CRITERIA } from './constants';
 
 export function generateLetterQualityCheckPrompt(letter: string, request: LetterRequest) {
   const initialInfo = request.initialQuestions;
